[fixup] Move Projection into o.a.f.table.connector
